Moved the queries to the queries package.
