+++ src/QueryParser/DatabaseHelper.java
@@ -2,6 +2,7 @@ package QueryParser;
 
 import Model.*;
 import common.Constants;
+import queries.*;
 
 import java.io.File;
 import java.util.ArrayList;
